// src/app/modules/admin/models/feedback.numbers.ts
export interface FeedbackNumbers {
  total: number;
  goodFeedbackTotal: number;
  sadFeedbackTotal: number;
  vgoodFeedbackTotal: number;
  vsadFeedbackTotal: number;
  vgoodFeedbackPercentage: number;
  goodFeedbackPercentage: number;
  vsadFeedbackPercentage: number;
  sadFeedbackPercentage: number;
}